Remove coupling to multi-platform lib
