--- data/src/main/java/com/dpanger/vehicles/data/MakeRepository.kt
@@ -1,11 +1,0 @@
-package com.dpanger.vehicles.data
-
-import uniffi.vehicles.Make
-import kotlin.coroutines.CoroutineContext
-
-interface MakeRepository {
-    suspend fun forManufacturer(id: String): List<Make>?
-}
-
-fun makeRepository(coroutineContext: CoroutineContext): MakeRepository =
-    MakeRepositoryImpl(coroutineContext)
