--- data/src/main/java/com/dpanger/vehicles/data/ManufacturerRepository.kt
@@ -1,10 +1,0 @@
-package com.dpanger.vehicles.data
-
-import uniffi.vehicles.Manufacturer
-
-interface ManufacturerRepository {
-    suspend fun all(): List<Manufacturer>?
-}
-
-fun manufacturerRepository(): ManufacturerRepository =
-    ManufacturerRepositoryImpl()
