+++ data/src/main/java/com/dpanger/vehicles/data/makes/Make.kt
@@ -1,0 +1,3 @@
+package com.dpanger.vehicles.data.makes
+
+data class Make(val name: String)
