+++ data/src/main/java/com/dpanger/vehicles/data/makes/MakeRepository.kt
@@ -1,0 +1,9 @@
+package com.dpanger.vehicles.data.makes
+
+import kotlin.coroutines.CoroutineContext
+
+interface MakeRepository {
+    suspend fun forManufacturer(id: String): List<Make>?
+}
+
+fun makeRepository(coroutineContext: CoroutineContext): MakeRepository = MakeRepositoryImpl(coroutineContext)
