+++ data/src/main/java/com/dpanger/vehicles/data/manufacturers/Manufacturer.kt
@@ -1,0 +1,6 @@
+package com.dpanger.vehicles.data.manufacturers
+
+data class Manufacturer(
+    val id: String,
+    val name: String,
+)
