+++ data/src/main/java/com/dpanger/vehicles/data/manufacturers/ManufacturerRepository.kt
@@ -1,0 +1,7 @@
+package com.dpanger.vehicles.data.manufacturers
+
+interface ManufacturerRepository {
+    suspend fun all(): List<Manufacturer>?
+}
+
+fun manufacturerRepository(): ManufacturerRepository = ManufacturerRepositoryImpl()
