+++ data/src/main/java/com/dpanger/vehicles/data/manufacturers/ManufacturerRepositoryImpl.kt
@@ -1,7 +1,6 @@
-package com.dpanger.vehicles.data
+package com.dpanger.vehicles.data.manufacturers
 
 import android.util.Log
-import uniffi.vehicles.Manufacturer
 import uniffi.vehicles.VehiclesException
 import uniffi.vehicles.manufacturers
 
@@ -9,6 +8,12 @@ internal class ManufacturerRepositoryImpl : ManufacturerRepository {
     override suspend fun all(): List<Manufacturer>? =
         try {
             manufacturers()
+                .map {
+                    Manufacturer(
+                        id = it.id,
+                        name = it.name,
+                    )
+                }
         } catch (exception: VehiclesException) {
             Log.e("Manufacturers", "Failed fetching manufacturers", exception)
             null
